requirements.txt
asgiref==3.2.3
asn1crypto==1.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-mathfilters==0.4.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-logging==1.14.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
mysqlclient==1.4.6
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8